move provider related code into provider package
